requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-cli
langchain-community>=0.0.10
langchain-together
python-dotenv>=1.0.0
python-telegram-bot>=20.7
streamlit>=1.29.0

# Vector store and embeddings
faiss-cpu
sentence-transformers>=2.2.2

# Document processing
openpyxl>=3.1.2
PyPDF2>=3.0.0
python-docx>=1.0.0

# LLM APIs
openai>=1.12.0

# Azure services
azure-ai-textanalytics>=5.3.0
azure-cognitiveservices-speech>=1.32.1

# Development tools
black>=23.12.1
pylint>=3.0.3

# LangChain migration
langchain-openai
langchain-text-splitters
langchain-core

ratelimit 
langchain-openai 
langchain-community
azure-ai-translation-text